# .spec/openapi.yml
openapi: 3.0.1
info:
  title: 'Kr.Carevo.UMR.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null'
  version: '1.0'
paths:
  /verify:
    get:
      tags:
        - Healthcheck
      summary: v1 health check.
      description: verify the health of the application.
      requestBody:
        content: { }
      responses:
        '200':
          description: OK
  /api/user/v1/register:
    post:
      tags:
        - User Registration
      summary: v1 user registration.
      description: user registration endpoint for UMR.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - { }
  '/api/user/v1/{userId}/skills':
    post:
      tags:
        - User Skills Update
      summary: v1 user skills update.
      description: user skills update endpoint for UMR.
      parameters:
        - name: userId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SkillDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SkillDto'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - { }
  '/api/user/v1/{id}':
    get:
      tags:
        - User Details
      summary: v1 user details.
      description: user details endpoint for UMR.
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content: { }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - { }
  '/api/user/v1/{userId}/employment':
    post:
      tags:
        - User Employment Update
      summary: v1 user employment update.
      description: user employment update endpoint for UMR.
      parameters:
        - name: userId
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/EmploymentDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security:
        - { }
components:
  schemas:
    AddressDto:
      type: object
      properties:
        line1:
          type: string
          nullable: true
        line2:
          type: string
          nullable: true
        suburb:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
      additionalProperties: false
    ContactDto:
      type: object
      properties:
        email:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        mobileNumber:
          type: string
          nullable: true
        isValid:
          type: boolean
          readOnly: true
      additionalProperties: false
    EmploymentDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        company:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        logo:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        userId:
          type: integer
          format: int32
        duration:
          type: string
          nullable: true
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectDto'
          nullable: true
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    ProjectDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        employmentId:
          type: integer
          format: int32
          nullable: true
        userId:
          type: integer
          format: int32
          nullable: true
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillDto'
          nullable: true
      additionalProperties: false
    ResponseDto:
      type: object
      properties:
        uri:
          type: string
          nullable: true
        responses:
          type: array
          items:
            $ref: '#/components/schemas/ResponseItemDto'
          nullable: true
      additionalProperties: false
    ResponseItemDto:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ResponseStatus'
        id:
          type: integer
          format: int32
        keyIdentifiers:
          type: array
          items:
            $ref: '#/components/schemas/StringObjectKeyValuePair'
          nullable: true
      additionalProperties: false
    ResponseStatus:
      enum:
        - Created
        - Failed
        - Skipped
      type: string
    SkillDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        code:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        effectiveDate:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    StringObjectKeyValuePair:
      type: object
      properties:
        key:
          type: string
          nullable: true
        value:
          nullable: true
      additionalProperties: false
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        dob:
          type: string
          format: date-time
        address:
          $ref: '#/components/schemas/AddressDto'
        contact:
          $ref: '#/components/schemas/ContactDto'
        name:
          type: string
          nullable: true
      additionalProperties: false
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        dob:
          type: string
          format: date-time
        name:
          type: string
          nullable: true
          readOnly: true
        contact:
          $ref: '#/components/schemas/ContactDto'
        status:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    Void:
      type: object
      additionalProperties: false